requirements.txt
fastapi
uvicorn[standard]
pydantic
langchain
langchain-openai
langchain-deepseek
python-dotenv
